GH Actions: change PHAR building to reusable workflow
The workflows currently contain two jobs which build the PHAR files.
This means that any changes to the steps in these jobs have to be made in two places.

With this in mind, it makes sense to change the PHAR building to a reusable workflow, which can then be used by both jobs.
With this change, any changes to the steps of the job will only need to be made in one place.

This commit makes it so.